License change to LGPLv3
Rationale:  I do not want developers to write nonfree software using
libleet, however I kind of need compatibility with GPLv2 and GPLv3
software, rather than just AGPLv3 licensed software.

One example of this is the Wt C++ library, which I plan on using
very shortly. In any case, with the LGPL is it still not permitted
to distribute **libleet** solely in binary form, source code must
still be provided and the license must still be LGPLv3. The only
notable difference is that you may now link programs with libleet
that are not licensed under the AGPLv3 software license.

In short; the pros outweigh the cons for me personally. Besides,
the library cannot be used directly to write nonfree software,
and there are other solutions right now that have more permissive
licensing available.